Exchange partial sigs in tx_signatures
Instead of exchanging partial sigs in `commit_sig` (which lets the user
provide complete witnesses in their `tx_signatures`), we now exchange
partial signatures to swap inputs directly in `tx_signatures`, and each
peer can independently create the complete witnesses.

This provides a cleaner separation of concerns between `commit_sig` and
`tx_signatures`.

+++ src/commonMain/kotlin/fr/acinq/lightning/transactions/Scripts.kt
@@ -33,7 +33,7 @@ object Scripts {
     /**
      * @return the script used for a 2-of-2 swap-in as used in Phoenix.
      */
-    fun swapIn2of2(userKey: PublicKey, serverKey: PublicKey, delayedRefund: Long): List<ScriptElt> {
+    fun swapIn2of2(userKey: PublicKey, serverKey: PublicKey, delayedRefund: Int): List<ScriptElt> {
         // This script was generated with https://bitcoin.sipa.be/miniscript/ using the following miniscript policy:
         // and(pk(<user_key>),or(99@pk(<server_key>),older(<delayed_refund>)))
         // @formatter:off
@@ -46,12 +46,12 @@ object Scripts {
         // @formatter:on
     }
 
-    fun witnessSwapIn2of2(userSig: ByteVector64, userKey: PublicKey, serverSig: ByteVector64, serverKey: PublicKey, delayedRefund: Long): ScriptWitness {
+    fun witnessSwapIn2of2(userSig: ByteVector64, userKey: PublicKey, serverSig: ByteVector64, serverKey: PublicKey, delayedRefund: Int): ScriptWitness {
         val redeemScript = swapIn2of2(userKey, serverKey, delayedRefund)
         return ScriptWitness(listOf(der(serverSig, SigHash.SIGHASH_ALL), der(userSig, SigHash.SIGHASH_ALL), Script.write(redeemScript).byteVector()))
     }
 
-    fun witnessSwapIn2of2Refund(userSig: ByteVector64, userKey: PublicKey, serverKey: PublicKey, delayedRefund: Long): ScriptWitness {
+    fun witnessSwapIn2of2Refund(userSig: ByteVector64, userKey: PublicKey, serverKey: PublicKey, delayedRefund: Int): ScriptWitness {
         val redeemScript = swapIn2of2(userKey, serverKey, delayedRefund)
         return ScriptWitness(listOf(ByteVector.empty, der(userSig, SigHash.SIGHASH_ALL), Script.write(redeemScript).byteVector()))
     }
